refactor: compound children에 해당하는 컴포넌트 Profile, Content, Footer 분리
